+++ tests/numerical/cdf/test_cases/frank/frank_cdf_test2.json
@@ -1,0 +1,57 @@
+{
+    "metadata": {
+        "test_type": "bivariate_pdf",
+        "dataset_type": "handcrafted",
+        "scripts": {
+            "output": {
+                "R": "bivariate_cdf.r",
+                "M": "bivariate_cdf.m"
+            }
+        },
+        "description": "Evaluate the Frank CDF with a negative theta.",
+        "parameters": {
+            "description": ""
+        }
+    },
+    "test": {
+        "class": "copulas.bivariate.frank.Frank",
+        "kwargs": {}
+    },
+    "input": {
+        "theta": -1.6,
+        "points": "bivariate_cdf_input_1.csv"
+    },
+    "output": {
+        "R": [
+            0.0401544009502728,
+            0.0721572538056803,
+            0.197672839962374,
+            0.280600356750086,
+            0.0635157669024075,
+            0.112533939103844,
+            0.293633597409801,
+            0.405589734543834,
+            0.0915641264874986,
+            0.159631010324871,
+            0.396055117904849,
+            0.533326754484268
+        ],
+        "M": [
+            0.0401544009502728,
+            0.0721572538056803,
+            0.197672839962374,
+            0.280600356750086,
+            0.0635157669024075,
+            0.112533939103844,
+            0.293633597409801,
+            0.405589734543834,
+            0.0915641264874986,
+            0.159631010324871,
+            0.396055117904849,
+            0.533326754484268
+        ]
+    },
+    "settings": {
+        "rtol": 1e-3
+    }
+}
