+++ tests/numerical/cdf/test_cases/gumbel/gumbel_cdf_test2.json
@@ -1,0 +1,57 @@
+{
+    "metadata": {
+        "test_type": "bivariate_pdf",
+        "dataset_type": "handcrafted",
+        "scripts": {
+            "output": {
+                "R": "bivariate_cdf.r",
+                "M": "bivariate_cdf.m"
+            }
+        },
+        "description": "Evaluate the Gumbel CDF with a large theta.",
+        "parameters": {
+            "description": ""
+        }
+    },
+    "test": {
+        "class": "copulas.bivariate.gumbel.Gumbel",
+        "kwargs": {}
+    },
+    "input": {
+        "theta": 3.4,
+        "points": "bivariate_cdf_input_1.csv"
+    },
+    "output": {
+        "R": [
+            0.177058780294955,
+            0.256822556715681,
+            0.32803564347916,
+            0.32996397978927,
+            0.193087656453177,
+            0.304330068566531,
+            0.46311415082736,
+            0.469871064963651,
+            0.198307626275289,
+            0.323315907621493,
+            0.587192372928239,
+            0.609538232535919
+        ],
+        "M": [
+            0.177058780294955,
+            0.256822556715681,
+            0.32803564347916,
+            0.32996397978927,
+            0.193087656453177,
+            0.304330068566531,
+            0.46311415082736,
+            0.469871064963651,
+            0.198307626275289,
+            0.323315907621493,
+            0.587192372928239,
+            0.609538232535919
+        ]
+    },
+    "settings": {
+        "rtol": 1e-3
+    }
+}
