--- tests/numerical/cdf/tests/clayton/clayton_cdf_test1.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 0.7,
-        "points": "bivariate_cdf_1.csv"
-    },
-    "output": {
-        "R": [
-            0.126218876493393,
-            0.178120287329811,
-            0.279006338667201,
-            0.31407676544624,
-            0.149538189799709,
-            0.221835511366905,
-            0.380341492470757,
-            0.441284137797374,
-            0.167113398502451,
-            0.257308993945402,
-            0.474821562888243,
-            0.565731674642723
-        ],
-        "M": [
-            0.126218876493393,
-            0.178120287329811,
-            0.279006338667201,
-            0.31407676544624,
-            0.149538189799709,
-            0.221835511366905,
-            0.380341492470757,
-            0.441284137797374,
-            0.167113398502451,
-            0.257308993945402,
-            0.474821562888243,
-            0.565731674642723
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_cdf.r",
-                "M": "bivariate_cdf.m"
-            }
-        },
-        "description": "Evaluate the Clayton CDF with a small theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
