--- tests/numerical/cdf/tests/clayton/clayton_cdf_test2.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 1.6,
-        "points": "bivariate_cdf_1.csv"
-    },
-    "output": {
-        "R": [
-            0.164081607303501,
-            0.226169514188503,
-            0.306780352172945,
-            0.323732449422615,
-            0.180468389237123,
-            0.267627150915129,
-            0.415488614967996,
-            0.454569699288105,
-            0.18931950202112,
-            0.293769182145534,
-            0.510195871025199,
-            0.58025841288278
-        ],
-        "M": [
-            0.164081607303501,
-            0.226169514188503,
-            0.306780352172945,
-            0.323732449422615,
-            0.180468389237123,
-            0.267627150915129,
-            0.415488614967996,
-            0.454569699288105,
-            0.18931950202112,
-            0.293769182145534,
-            0.510195871025199,
-            0.58025841288278
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_cdf.r",
-                "M": "bivariate_cdf.m"
-            }
-        },
-        "description": "Evaluate the Clayton CDF with a medium theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
