--- tests/numerical/cdf/tests/clayton/clayton_cdf_test3.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 3.4,
-        "points": "bivariate_cdf_1.csv"
-    },
-    "output": {
-        "R": [
-            0.190592098776929,
-            0.270058178254211,
-            0.325221843599552,
-            0.329041776940273,
-            0.197120359803162,
-            0.307071738905805,
-            0.448880920836403,
-            0.465524347625505,
-            0.198932669671263,
-            0.320812128352332,
-            0.551304135589609,
-            0.596308571630797
-        ],
-        "M": [
-            0.190592098776929,
-            0.270058178254211,
-            0.325221843599552,
-            0.329041776940273,
-            0.197120359803162,
-            0.307071738905805,
-            0.448880920836403,
-            0.465524347625505,
-            0.198932669671263,
-            0.320812128352332,
-            0.551304135589609,
-            0.596308571630797
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_cdf.r",
-                "M": "bivariate_cdf.m"
-            }
-        },
-        "description": "Evaluate the Clayton CDF with a large theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
