--- tests/numerical/cdf/tests/frank/frank_cdf_test1.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.frank.Frank",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 0.7,
-        "points": "bivariate_cdf_1.csv"
-    },
-    "output": {
-        "R": [
-            0.0785896492359957,
-            0.126136390235633,
-            0.249870288875225,
-            0.303691865497178,
-            0.10791185809841,
-            0.174313227857916,
-            0.351491790349644,
-            0.430738251832419,
-            0.135027233031875,
-            0.219442968427702,
-            0.450324183644844,
-            0.556585812296097
-        ],
-        "M": [
-            0.0785896492359957,
-            0.126136390235633,
-            0.249870288875225,
-            0.303691865497178,
-            0.10791185809841,
-            0.174313227857916,
-            0.351491790349644,
-            0.430738251832419,
-            0.135027233031875,
-            0.219442968427702,
-            0.450324183644844,
-            0.556585812296097
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_cdf.r",
-                "M": "bivariate_cdf.m"
-            }
-        },
-        "description": "Evaluate the Frank CDF with a small theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
