+++ tests/numerical/fit/test_cases/frank/frank_fit_test2.json
@@ -1,0 +1,35 @@
+{
+    "metadata": {
+        "test_type": "bivariate_fit",
+        "dataset_type": "handcrafted",
+        "scripts": {
+            "input": "bivariate_fit_dataset_2.py",
+            "output": {
+                "R": "bivariate_fit.r",
+                "M": "bivariate_fit.m"
+            }
+        },
+        "description": "Test Frank on a weakly correlated dataset.",
+        "parameters": {
+            "description": ""
+        }
+    },
+    "test": {
+        "class": "copulas.bivariate.frank.Frank",
+        "kwargs": {}
+    },
+    "input": "bivariate_fit_2.csv",
+    "output": {
+        "R": {
+            "theta": 1.30743452080717,
+            "tau": 0.142857142857143
+        },
+        "M": {
+            "theta": 1.30743452080717,
+            "tau": 0.142857142857143
+        }
+    },
+    "settings": {
+        "rtol": 1e-5
+    }
+}
