+++ tests/numerical/fit/test_cases/gumbel/gumbel_fit_test1.json
@@ -1,0 +1,35 @@
+{
+    "metadata": {
+        "test_type": "bivariate_fit",
+        "dataset_type": "handcrafted",
+        "scripts": {
+            "input": "bivariate_fit_dataset_1.py",
+            "output": {
+                "R": "bivariate_fit.r",
+                "M": "bivariate_fit.m"
+            }
+        },
+        "description": "Test Gumbel on a highly correlated dataset.",
+        "parameters": {
+            "description": ""
+        }
+    },
+    "test": {
+        "class": "copulas.bivariate.gumbel.Gumbel",
+        "kwargs": {}
+    },
+    "input": "bivariate_fit_1.csv",
+    "output": {
+        "R": {
+            "theta": 4.71090054704673,
+            "tau": 0.787726361443376
+        },
+        "M": {
+            "theta": 4.71090054704673,
+            "tau": 0.787726361443376
+        }
+    },
+    "settings": {
+        "rtol": 1e-7
+    }
+}
