--- tests/numerical/fit/tests/clayton/clayton_fit_test2.json
@@ -1,35 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": "bivariate_fit_2.csv",
-    "output": {
-        "R": {
-            "theta": 0.333333333333333,
-            "tau": 0.142857142857143
-        },
-        "M": {
-            "theta": 0.333333333333333,
-            "tau": 0.142857142857143
-        }
-    },
-    "settings": {
-        "rtol": 1e-7
-    },
-    "metadata": {
-        "test_type": "bivariate_fit",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "input": "bivariate_fit_dataset_2.py",
-            "output": {
-                "R": "bivariate_fit.r",
-                "M": "bivariate_fit.m"
-            }
-        },
-        "description": "Test Clayton on a weakly correlated dataset.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
