--- tests/numerical/fit/tests/gumbel/gumbel_fit_test2.json
@@ -1,35 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.gumbel.Gumbel",
-        "kwargs": {}
-    },
-    "input": "bivariate_fit_2.csv",
-    "output": {
-        "R": {
-            "theta": 1.16666666666667,
-            "tau": 0.142857142857143
-        },
-        "M": {
-            "theta": 1.16666666666667,
-            "tau": 0.142857142857143
-        }
-    },
-    "settings": {
-        "rtol": 1e-7
-    },
-    "metadata": {
-        "test_type": "bivariate_fit",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "input": "bivariate_fit_dataset_2.py",
-            "output": {
-                "R": "bivariate_fit.r",
-                "M": "bivariate_fit.m"
-            }
-        },
-        "description": "Test Gumbel on a weakly correlated dataset.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
