+++ tests/numerical/pdf/test_cases/frank/frank_pdf_test2.json
@@ -1,0 +1,57 @@
+{
+    "metadata": {
+        "test_type": "bivariate_pdf",
+        "dataset_type": "handcrafted",
+        "scripts": {
+            "output": {
+                "R": "bivariate_pdf.r",
+                "M": "bivariate_pdf.m"
+            }
+        },
+        "description": "Evaluate the Frank PDF with a negative theta.",
+        "parameters": {
+            "description": ""
+        }
+    },
+    "test": {
+        "class": "copulas.bivariate.frank.Frank",
+        "kwargs": {}
+    },
+    "input": {
+        "theta": -1.6,
+        "points": "bivariate_pdf_input_1.csv"
+    },
+    "output": {
+        "R": [
+            0.831126357814658,
+            0.923686482388915,
+            1.13538017932813,
+            1.18010037791185,
+            0.964900963313247,
+            1.01553141529965,
+            1.04487068440181,
+            0.989687038999556,
+            1.10353144027994,
+            1.09275432149173,
+            0.941900407225639,
+            0.822731833301323
+        ],
+        "M": [
+            0.831126357814658,
+            0.923686482388915,
+            1.13538017932813,
+            1.18010037791185,
+            0.964900963313247,
+            1.01553141529965,
+            1.04487068440181,
+            0.989687038999556,
+            1.10353144027994,
+            1.09275432149173,
+            0.941900407225639,
+            0.822731833301323
+        ]
+    },
+    "settings": {
+        "rtol": 1e-3
+    }
+}
