--- tests/numerical/pdf/tests/clayton/clayton_pdf_test1.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 0.7,
-        "points": "bivariate_pdf_1.csv"
-    },
-    "output": {
-        "R": [
-            1.20205011731765,
-            1.17277217836311,
-            0.936073172626286,
-            0.83109161504681,
-            0.989775401081488,
-            1.08863785272935,
-            1.07934495343816,
-            1.03037249519005,
-            0.829591436597552,
-            0.99771482232722,
-            1.18010879940507,
-            1.20071828403312
-        ],
-        "M": [
-            1.20205011731765,
-            1.17277217836311,
-            0.936073172626286,
-            0.83109161504681,
-            0.989775401081488,
-            1.08863785272935,
-            1.07934495343816,
-            1.03037249519005,
-            0.829591436597552,
-            0.99771482232722,
-            1.18010879940507,
-            1.20071828403312
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_pdf.r",
-                "M": "bivariate_pdf.m"
-            }
-        },
-        "description": "Evaluate the Clayton PDF with a small theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
