--- tests/numerical/pdf/tests/clayton/clayton_pdf_test2.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.clayton.Clayton",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 1.6	,
-        "points": "bivariate_pdf_1.csv"
-    },
-    "output": {
-        "R": [
-            1.53960992606222,
-            1.61184744189229,
-            0.791117647796511,
-            0.535289542840989,
-            0.91564204219723,
-            1.30319495710224,
-            1.1277051091007,
-            0.887994798793683,
-            0.568405501434614,
-            0.978601138839519,
-            1.35624833021077,
-            1.25688436755159
-        ],
-        "M": [
-            1.53960992606222,
-            1.61184744189229,
-            0.791117647796511,
-            0.535289542840989,
-            0.91564204219723,
-            1.30319495710224,
-            1.1277051091007,
-            0.887994798793683,
-            0.568405501434614,
-            0.978601138839519,
-            1.35624833021077,
-            1.25688436755159
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_pdf.r",
-                "M": "bivariate_pdf.m"
-            }
-        },
-        "description": "Evaluate the Clayton PDF with a medium theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
