--- tests/numerical/pdf/tests/gumbel/gumbel_pdf_test1.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.gumbel.Gumbel",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 1.6,
-        "points": "bivariate_pdf_1.csv"
-    },
-    "output": {
-        "R": [
-            1.3466603461109,
-            1.33436989228653,
-            0.844034579550222,
-            0.401584644502942,
-            1.10035364987839,
-            1.22397629649902,
-            1.07651272051626,
-            0.574400583179372,
-            0.840259033271954,
-            1.02839689426663,
-            1.33135685092464,
-            0.868366681780188
-        ],
-        "M": [
-            1.3466603461109,
-            1.33436989228653,
-            0.844034579550222,
-            0.401584644502942,
-            1.10035364987839,
-            1.22397629649902,
-            1.07651272051626,
-            0.574400583179372,
-            0.840259033271954,
-            1.02839689426663,
-            1.33135685092464,
-            0.868366681780188
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_pdf.r",
-                "M": "bivariate_pdf.m"
-            }
-        },
-        "description": "Evaluate the Gumbel PDF with a small theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
