--- tests/numerical/pdf/tests/gumbel/gumbel_pdf_test2.json
@@ -1,57 +1,0 @@
-{
-    "test": {
-        "class": "copulas.bivariate.gumbel.Gumbel",
-        "kwargs": {}
-    },
-    "input": {
-        "theta": 3.4,
-        "points": "bivariate_pdf_1.csv"
-    },
-    "output": {
-        "R": [
-            1.84362775082523,
-            2.45131531246008,
-            0.256648173862043,
-            0.0123801570777976,
-            0.740387192158614,
-            1.67877054715316,
-            0.781068175711491,
-            0.0410323620974807,
-            0.231523459171766,
-            0.661658756591641,
-            2.16744242637305,
-            0.157797071824743
-        ],
-        "M": [
-            1.84362775082523,
-            2.45131531246008,
-            0.256648173862043,
-            0.0123801570777976,
-            0.740387192158614,
-            1.67877054715316,
-            0.781068175711491,
-            0.0410323620974807,
-            0.231523459171766,
-            0.661658756591641,
-            2.16744242637305,
-            0.157797071824743
-        ]
-    },
-    "settings": {
-        "rtol": 1e-3
-    },
-    "metadata": {
-        "test_type": "bivariate_pdf",
-        "dataset_type": "handcrafted",
-        "scripts": {
-            "output": {
-                "R": "bivariate_pdf.r",
-                "M": "bivariate_pdf.m"
-            }
-        },
-        "description": "Evaluate the Gumbel PDF with a large theta.",
-        "parameters": {
-            "description": ""
-        }
-    }
-}
